Implement unlimited storage
This commit makes separate changes to the storage quota enforcer and the quota
manager.  It seems likely that only one of these two files actually needs to
change, and this should be checked at a later date.

The approach taken is brutal, but apparently effective:

  1. anywhere where the IGNORE_QUOTA option was previously checked, we hardcode
     option-value == true
  2. instead of reporting the true value of used memory from the
     StorageQuotaEnforcer, we have hard-coded the relevant function to return
     the constant value zero

Neither of these changes should be considered long-term, but either or both of
them is likely to point the direction toward where more constructive changes
could be made, and potentially accepted upstream.